Correct import statements
